requirements.txt
annotated-types==0.7.0
anyio==4.8.0
av==11.0.0
blis==1.1.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
ctranslate2==4.3.1
cymem==2.0.10
distro==1.9.0
enum34==1.1.10
exceptiongroup==1.2.2
faster-whisper==1.0.1
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
h11==0.14.0
halo==0.0.31
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
Jinja2==3.1.4
jiter==0.8.2
langcodes==3.5.0
language_data==1.3.0
log-symbols==0.0.14
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.11
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.19.0
openai==1.59.6
opencv-python==4.10.0.84
packaging==24.1
preshed==3.0.9
protobuf==5.27.4
pvporcupine==1.9.5
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pygame==2.6.0
Pygments==2.19.1
pyreadline3==3.4.1
python-dotenv==1.0.1
pyttsx3==2.98
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scipy==1.10.1
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spinners==0.0.24
srsly==2.5.0
srt==3.5.3
sympy==1.13.2
termcolor==2.4.0
thinc==8.3.3
tokenizers==0.15.2
torch==2.2.2
torchaudio==2.2.2
tqdm==4.66.5
triton==2.2.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.2
vosk==0.3.45
wasabi==1.1.3
weasel==0.4.1
webrtcvad==2.0.10
websockets==13.0
word2number==1.1
wrapt==1.17.1
